+++ src/main/java/com/saiko/bidmarket/product/repository/dto/UserProductSelectQueryParameter.java
@@ -14,13 +14,13 @@ public class UserProductSelectQueryParameter {
 
   private final long userId;
 
-  private final int offset;
+  private final long offset;
 
   private final int limit;
 
   private final Sort sort;
 
-  private UserProductSelectQueryParameter(long userId, int offset, int limit, Sort sort) {
+  private UserProductSelectQueryParameter(long userId, long offset, int limit, Sort sort) {
     Assert.isTrue(userId > 0, "User id must be positive");
     Assert.isTrue(offset >= 0, "offset must not be negative");
     Assert.isTrue(limit > 0, "limit must be positive");
